Main: do not include include/VBox/settings.h from other header files but only from implementations that need it (save compile time)
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@16560 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/ApplianceImpl.cpp
@@ -1564,7 +1564,7 @@ DECLCALLBACK(int) Appliance::taskThread(RTTHREAD aThread, void *pvUser)
 
     /* For now we report 2 steps for every virtual system. Later we may add the
        progress of the image cloning. */
-    float opCountMax = 100.0/(app->m->llVirtualSystems.size() * 2);
+    float opCountMax = (float)100.0 / (app->m->llVirtualSystems.size() * 2);
     uint32_t opCount = 0;
 
     list<VirtualSystem>::const_iterator it;
@@ -1612,7 +1612,7 @@ DECLCALLBACK(int) Appliance::taskThread(RTTHREAD aThread, void *pvUser)
             std::list<VirtualSystemDescriptionEntry*> vsdeRAM = vsdescThis->findByType(VirtualSystemDescriptionType_Memory);
             Assert(vsdeRAM.size() == 1);
             const Utf8Str &memoryVBox = vsdeRAM.front()->strConfig;
-            uint64_t tt = RTStrToUInt64(memoryVBox.c_str()) / _1M;
+            ULONG tt = (ULONG)RTStrToUInt64(memoryVBox.c_str()) / _1M;
 
             rc = pNewMachine->COMSETTER(MemorySize)(tt);
             CheckComRCThrowRC(rc);
